Deleted extraneous file.
