+++ src/test/java/org/broadinstitute/hellbender/tools/copynumber/utils/annotatedregion/SimpleAnnotatedGenomicRegionUnitTest.java
@@ -1,15 +1,16 @@
 package org.broadinstitute.hellbender.tools.copynumber.utils.annotatedregion;
 
 import com.google.common.collect.ImmutableSortedMap;
-import com.google.common.collect.Lists;
 import org.broadinstitute.hellbender.utils.SimpleInterval;
 import org.broadinstitute.hellbender.utils.test.BaseTest;
 import org.testng.Assert;
+import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 import org.testng.collections.Sets;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Set;
 
@@ -18,7 +19,7 @@ public class SimpleAnnotatedGenomicRegionUnitTest extends BaseTest {
 
     @Test
     public void basicTest() throws IOException {
-        final Set<String> headersOfInterest = Sets.newHashSet(Lists.newArrayList("name", "learning_SAMPLE_0"));
+        final Set<String> headersOfInterest = Sets.newHashSet(Arrays.asList("name", "learning_SAMPLE_0"));
         final List<SimpleAnnotatedGenomicRegion> simpleAnnotatedGenomicRegions =
                 SimpleAnnotatedGenomicRegion.readAnnotatedRegions(new File(TEST_FILE), headersOfInterest);
 
@@ -29,7 +30,7 @@ public class SimpleAnnotatedGenomicRegionUnitTest extends BaseTest {
         Assert.assertTrue(simpleAnnotatedGenomicRegions.stream().allMatch(s -> s.getAnnotations().keySet().containsAll(headersOfInterest)));
 
         // Grab the first 15 and test values
-        List<SimpleAnnotatedGenomicRegion> gtRegions = Lists.newArrayList(
+        List<SimpleAnnotatedGenomicRegion> gtRegions = Arrays.asList(
                 new SimpleAnnotatedGenomicRegion(new SimpleInterval("1", 30365, 30503), ImmutableSortedMap.of("name", "target_1_None", "learning_SAMPLE_0", "2")),
                 new SimpleAnnotatedGenomicRegion(new SimpleInterval("1", 69088, 70010), ImmutableSortedMap.of("name", "target_2_None", "learning_SAMPLE_0", "2")),
                 new SimpleAnnotatedGenomicRegion(new SimpleInterval("1", 367656, 368599), ImmutableSortedMap.of("name", "target_3_None", "learning_SAMPLE_0", "2")),
@@ -49,5 +50,49 @@ public class SimpleAnnotatedGenomicRegionUnitTest extends BaseTest {
 
         Assert.assertEquals(simpleAnnotatedGenomicRegions.subList(0, gtRegions.size()), gtRegions);
 
+    }
+
+    @Test(dataProvider = "equalsTests")
+    public void testEquals(SimpleAnnotatedGenomicRegion test1, SimpleAnnotatedGenomicRegion test2, boolean gt) {
+        Assert.assertEquals(test1.equals(test2), gt);
+    }
+
+    @DataProvider(name = "equalsTests")
+    public Object [][] createEqualsTests() {
+
+        return new Object[][] {
+
+                {
+                        new SimpleAnnotatedGenomicRegion( new SimpleInterval("1", 100, 200),
+                                ImmutableSortedMap.of("Foo", "bar", "Foo1", "bar1")),
+                        new SimpleAnnotatedGenomicRegion( new SimpleInterval("1", 100, 200),
+                                ImmutableSortedMap.of("Foo", "bar", "Foo1", "bar1")),
+                        true
+                }, {
+                new SimpleAnnotatedGenomicRegion( new SimpleInterval("1", 100, 200),
+                        ImmutableSortedMap.of("Foo", "bar", "Foo1", "bar2")),
+                new SimpleAnnotatedGenomicRegion( new SimpleInterval("1", 100, 200),
+                        ImmutableSortedMap.of("Foo", "bar", "Foo1", "bar1")),
+                false
+        }, {
+                new SimpleAnnotatedGenomicRegion( new SimpleInterval("1", 100, 200),
+                        ImmutableSortedMap.of("Foo", "bar", "Foo1", "bar1")),
+                new SimpleAnnotatedGenomicRegion( new SimpleInterval("1", 100, 200),
+                        ImmutableSortedMap.of("Foo", "bar", "Foobar1", "bar1")),
+                false
+        }, {
+                new SimpleAnnotatedGenomicRegion( new SimpleInterval("1", 1000, 2000),
+                        ImmutableSortedMap.of("Foo", "bar", "Foo1", "bar1")),
+                new SimpleAnnotatedGenomicRegion( new SimpleInterval("1", 100, 200),
+                        ImmutableSortedMap.of("Foo", "bar", "Foo1", "bar1")),
+                false
+        }, {
+                new SimpleAnnotatedGenomicRegion( new SimpleInterval("1", 100, 200),
+                        ImmutableSortedMap.of("Foo", "bar", "Foo1", "bar1")),
+                new SimpleAnnotatedGenomicRegion( new SimpleInterval("1", 100, 200),
+                        ImmutableSortedMap.of("Foo", "bar")),
+                false
+        }
+        };
     }
 }
